pyproject.toml
[project]
name = "vdes1000"
version = "0.1.0"
description = "A Python package to facilitate communication with the CML Microcircuits VDES1000 transceiver."
authors = [
    {name = "Jan Safar", email = "jan.safar@gla-rad.org"},
]
dependencies = [
    "pyyaml>=6.0.1",
    "bitstring>=4.1.4",
    "iec-61162 @ git+https://github.com/jan-safar/py_iec_61162.git@v0.1.0",
    "iec-62320 @ git+https://github.com/jan-safar/py_iec_62320.git@v0.1.0",
    "iec-pas-63343 @ git+https://github.com/jan-safar/py_iec_pas_63343.git@v0.1.0",
    "rec-itu-r-m-1371 @ git+https://github.com/jan-safar/py_rec_itu_r_m_1371.git@v0.1.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Apache-2.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
spyder = [
    "spyder-kernels==2.5.*",
]
dev = [
]
